// condoplanner-app/src/apiClient/models/EStatusEncomenda.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const EStatusEncomenda = {
    Recebido: 'RECEBIDO',
    Notificado: 'NOTIFICADO',
    AguardandoRetirada: 'AGUARDANDO_RETIRADA',
    Retirado: 'RETIRADO'
} as const;
export type EStatusEncomenda = typeof EStatusEncomenda[keyof typeof EStatusEncomenda];


export function instanceOfEStatusEncomenda(value: any): boolean {
    for (const key in EStatusEncomenda) {
        if (Object.prototype.hasOwnProperty.call(EStatusEncomenda, key)) {
            if (EStatusEncomenda[key as keyof typeof EStatusEncomenda] === value) {
                return true;
            }
        }
    }
    return false;
}

export function EStatusEncomendaFromJSON(json: any): EStatusEncomenda {
    return EStatusEncomendaFromJSONTyped(json, false);
}

export function EStatusEncomendaFromJSONTyped(json: any, ignoreDiscriminator: boolean): EStatusEncomenda {
    return json as EStatusEncomenda;
}

export function EStatusEncomendaToJSON(value?: EStatusEncomenda | null): any {
    return value as any;
}

export function EStatusEncomendaToJSONTyped(value: any, ignoreDiscriminator: boolean): EStatusEncomenda {
    return value as EStatusEncomenda;
}