// condoplanner-app/src/apiClient/models/PackageDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { EStatusEncomenda } from './EStatusEncomenda';
import {
    EStatusEncomendaFromJSON,
    EStatusEncomendaFromJSONTyped,
    EStatusEncomendaToJSON,
    EStatusEncomendaToJSONTyped,
} from './EStatusEncomenda';

/**
 * 
 * @export
 * @interface PackageDto
 */
export interface PackageDto {
    /**
     * 
     * @type {number}
     * @memberof PackageDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof PackageDto
     */
    carrier?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof PackageDto
     */
    receivedAt?: Date;
    /**
     * 
     * @type {EStatusEncomenda}
     * @memberof PackageDto
     */
    status?: EStatusEncomenda;
    /**
     * 
     * @type {string}
     * @memberof PackageDto
     */
    notes?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof PackageDto
     */
    pickedUpAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof PackageDto
     */
    pickupPersonName?: string | null;
    /**
     * 
     * @type {number}
     * @memberof PackageDto
     */
    condominiumId?: number;
    /**
     * 
     * @type {number}
     * @memberof PackageDto
     */
    residentialLinkId?: number;
}



/**
 * Check if a given object implements the PackageDto interface.
 */
export function instanceOfPackageDto(value: object): value is PackageDto {
    return true;
}

export function PackageDtoFromJSON(json: any): PackageDto {
    return PackageDtoFromJSONTyped(json, false);
}

export function PackageDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PackageDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'carrier': json['carrier'] == null ? undefined : json['carrier'],
        'receivedAt': json['receivedAt'] == null ? undefined : (new Date(json['receivedAt'])),
        'status': json['status'] == null ? undefined : EStatusEncomendaFromJSON(json['status']),
        'notes': json['notes'] == null ? undefined : json['notes'],
        'pickedUpAt': json['pickedUpAt'] == null ? undefined : (new Date(json['pickedUpAt'])),
        'pickupPersonName': json['pickupPersonName'] == null ? undefined : json['pickupPersonName'],
        'condominiumId': json['condominiumId'] == null ? undefined : json['condominiumId'],
        'residentialLinkId': json['residentialLinkId'] == null ? undefined : json['residentialLinkId'],
    };
}

export function PackageDtoToJSON(json: any): PackageDto {
    return PackageDtoToJSONTyped(json, false);
}

export function PackageDtoToJSONTyped(value?: PackageDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'carrier': value['carrier'],
        'receivedAt': value['receivedAt'] == null ? value['receivedAt'] : value['receivedAt'].toISOString(),
        'status': EStatusEncomendaToJSON(value['status']),
        'notes': value['notes'],
        'pickedUpAt': value['pickedUpAt'] == null ? value['pickedUpAt'] : value['pickedUpAt'].toISOString(),
        'pickupPersonName': value['pickupPersonName'],
        'condominiumId': value['condominiumId'],
        'residentialLinkId': value['residentialLinkId'],
    };
}