// condoplanner-app/src/apiClient/models/CreatePackageInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreatePackageInput
 */
export interface CreatePackageInput {
    /**
     * 
     * @type {string}
     * @memberof CreatePackageInput
     */
    carrier?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof CreatePackageInput
     */
    receivedAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof CreatePackageInput
     */
    notes?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CreatePackageInput
     */
    condominiumId?: number;
    /**
     * 
     * @type {number}
     * @memberof CreatePackageInput
     */
    residentialLinkId?: number;
}

/**
 * Check if a given object implements the CreatePackageInput interface.
 */
export function instanceOfCreatePackageInput(value: object): value is CreatePackageInput {
    return true;
}

export function CreatePackageInputFromJSON(json: any): CreatePackageInput {
    return CreatePackageInputFromJSONTyped(json, false);
}

export function CreatePackageInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreatePackageInput {
    if (json == null) {
        return json;
    }
    return {
        
        'carrier': json['carrier'] == null ? undefined : json['carrier'],
        'receivedAt': json['receivedAt'] == null ? undefined : (new Date(json['receivedAt'])),
        'notes': json['notes'] == null ? undefined : json['notes'],
        'condominiumId': json['condominiumId'] == null ? undefined : json['condominiumId'],
        'residentialLinkId': json['residentialLinkId'] == null ? undefined : json['residentialLinkId'],
    };
}

export function CreatePackageInputToJSON(json: any): CreatePackageInput {
    return CreatePackageInputToJSONTyped(json, false);
}

export function CreatePackageInputToJSONTyped(value?: CreatePackageInput | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'carrier': value['carrier'],
        'receivedAt': value['receivedAt'] == null ? value['receivedAt'] : value['receivedAt'].toISOString(),
        'notes': value['notes'],
        'condominiumId': value['condominiumId'],
        'residentialLinkId': value['residentialLinkId'],
    };
}