// condoplanner-app/src/apiClient/models/UpdatePackageInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdatePackageInput
 */
export interface UpdatePackageInput {
    /**
     * 
     * @type {string}
     * @memberof UpdatePackageInput
     */
    carrier?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdatePackageInput
     */
    notes?: string | null;
}

/**
 * Check if a given object implements the UpdatePackageInput interface.
 */
export function instanceOfUpdatePackageInput(value: object): value is UpdatePackageInput {
    return true;
}

export function UpdatePackageInputFromJSON(json: any): UpdatePackageInput {
    return UpdatePackageInputFromJSONTyped(json, false);
}

export function UpdatePackageInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdatePackageInput {
    if (json == null) {
        return json;
    }
    return {
        
        'carrier': json['carrier'] == null ? undefined : json['carrier'],
        'notes': json['notes'] == null ? undefined : json['notes'],
    };
}

export function UpdatePackageInputToJSON(json: any): UpdatePackageInput {
    return UpdatePackageInputToJSONTyped(json, false);
}

export function UpdatePackageInputToJSONTyped(value?: UpdatePackageInput | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'carrier': value['carrier'],
        'notes': value['notes'],
    };
}